A Java based Webserver
Java Based Web Server, being very lightweight and highly customizable.

Capable of a lot of stuff other servers can do plus being able to cache basically everything in RAM natively, which makes it extremely fast and easy to add lots of clever logic to it.

Not to be confused with actual professional webserver software in terms of security and reliability, but great for prototyping.